bug(CART-CPC-1559): Create v1 endpoints and change API calls to v1 (#1231)
**JIRA:**
Link:
https://champlainsaintlambert.atlassian.net/browse/CPC-1559?atlOrigin=eyJpIjoiZjlhM2NhNDdlMjM4NGY4OTk4MTA4Y2UzMjBmNmNkNWYiLCJwIjoiaiJ9
## Context:
There was no v1 controller in the API gateway for the cart-service. To
conform to best practices in docs, CartControllerV1 was created and
calls in the frontend now use the v1 controller.
## Does this PR change the .vscode folder in petclinic-frontend?:
No
## Changes
- CartControllerV1 was created
- CartControllerV1UnitTest was created
- fields in CartServiceClient have been updated to use lower camel case
- V1 package in presentation layer in tests in api-gateway was changed
to v1.
- API calls using the axiosInstance relating to cart functionality in
the frontend now declares useV2 as false.
## Does this use the v2 API?:
It does not. It actually creates v1 API and uses that instead.
## Does this add a new communication between services?:
No
## Before and After UI (Required for UI-impacting PRs)
No changes to UI