bug(VETS-CPC-1629): Fix the API calls on the vet details page (#1223)
**JIRA:** [link to jira
ticket](https://champlainsaintlambert.atlassian.net/browse/CPC-1629)
## Context:
There used to be a bug with the API calls where certain requests to the
backend were using the v1 API instead of the v2, which was breaking the
requests.

## Does this PR change the .vscode folder in petclinic-frontend?:
No

## Changes
- Changed certain axiosInstance API requests to use the correct and
working version of the API (v2 instead of v1)
- Brought back certain permissions that were commented out in the
previous code, which fixes every permission-based request to the API
(e.g. load the vet profile picture).

## Does this use the v2 API?:
Yes, it uses the v2 API because there are no existing endpoints
currently that work in the same way on the v1 API.

## Does this add a new communication between services?:
No

## Before and After UI (Required for UI-impacting PRs)
No changes made to the UI