bug(VIST-CPC-1647): fix post emergency visit (#1232)
**JIRA:** [link to jira
ticket](https://champlainsaintlambert.atlassian.net/browse/CPC-1647?atlOrigin=eyJpIjoiNDdiN2U0NTRmYzhhNGJhMDlkZDUzNTYzMjQ3YTU2YWUiLCJwIjoiaiJ9)
## Context:
What is the ticket about and why are we doing this change.

This ticket fixes an issue where users get an error when trying to add
an emergency visit. The change makes sure that the “Add Emergency Visit”
now works as intended.

## Does this PR change the .vscode folder in petclinic-frontend?:
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.

No.
<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.

- Redirects users to the correct page after adding an emergency visit.
- Ensures proper formatting of the date field.

## Does this use the v2 API?:
If the PR uses the v2 API, explain why

No, it does not.
## Does this add a new communication between services?:
If the PR adds a new communication between services, make sure to update
the C4 L2.
No.
## Before and After UI (Required for UI-impacting PRs)
If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.

N/A
## Dev notes (Optional)
Specific technical changes that should be noted

N/A
## Linked pull requests (Optional)
Pull request links

N/A