Bug/CUST-CPC-1588-Fix API gateway test coverage (#1244)
**JIRA:** link to jira ticket 
https://champlainsaintlambert.atlassian.net/browse/CPC-1588
## Context:
What is the ticket about and why are we doing this change.

The ticket is about increasing the jacoco code coverage for owner
related tests in the API gateway from around 60% to 80%+. We are doing
this change to achieve the minimum requirement of 80%+ jacoco code
coverage for the API gateway and get the 2-4 pts attributed to it in the
Sprint presentation.
## Does this PR change the .vscode folder in petclinic-frontend?:
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.
NA
<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.

The various changes are the addition of new folders to separate owner
related tests from other v1 tests. There is also the addition of many
files for pets, petTypes and Owner controller Unit and Integration
tests. Finally, commented out tests and code related to owners and pets
have also been removed. These changes affect no other modules than the
API gateway.
## Does this use the v2 API?:
If the PR uses the v2 API, explain why
NA
## Does this add a new communication between services?:
If the PR adds a new communication between services, make sure to update
the C4 L2.
NA
## Before and After UI (Required for UI-impacting PRs)
If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.
NA
## Dev notes (Optional)
Specific technical changes that should be noted

Someone previously brought back into main the old endpoints that were
modified by the owner team in Sprint 1. This has caused the resurgence
of incorrect endpoints such as /pet, /owner and
/pet/owner/{ownerId}/pets/{petId} among the current ones we have. Seeing
as it would take an entire ticket to refactor back all these changes
across tests and API src code related to owners, they endpoints were
used for these tests knowing they did not match proper coding standards
and would have to be changed once again down the line.
## Linked pull requests (Optional)
Pull request links