feat/CUST-CPC-1452 Customer Photo Get (#1266)
**JIRA:** link to jira ticket
https://champlainsaintlambert.atlassian.net/browse/CPC-1452

## Context:
What is the ticket about and why are we doing this change.

The ticket is about adding the ability to display a profile picture for
owners in their profile using the file system. The ticket also asked for
proper http statuses to be return which is the case. This change is
being done to remove the need for different picture databases in each
service that wants pictures, and instead uses a commonly shared file
system.

## Does this PR change the .vscode folder in petclinic-frontend?:
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.
NA
<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>
NA
## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.

The changes made are the deletion of the obsolete Photo classes inside
the customer service reactive files, the addition of methods to handle
the GET owner picture feature, in both the API gateway and the customer
service.

## Does this use the v2 API?:
If the PR uses the v2 API, explain why

No
## Does this add a new communication between services?:
If the PR adds a new communication between services, make sure to update
the C4 L2.

NA
## Before and After UI (Required for UI-impacting PRs)
If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.
<img width="1886" height="894" alt="Capture d’écran 2025-10-11 151014"
src="https://github.com/user-attachments/assets/ae89df11-d0ce-40b7-bae7-3fcd7abf5daa"
/>

<img width="1919" height="1078" alt="Capture d’écran 2025-10-11 150448"
src="https://github.com/user-attachments/assets/283aeff7-56eb-4104-8397-44613e9b32c8"
/>

## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links