feat(AUTH-CPC-1686): Centralize front-end data (#1278)
**JIRA:**
[link](https://champlainsaintlambert.atlassian.net/browse/CPC-1686)

## Context:

This tickets goal is to add a centralized area for static data that is
not important enough to be an env and used in multiple tsx files. This
allows for simplicity for front-end designers by adding a single source
of truth for front-end static data.

## Does this PR change the .vscode folder in petclinic-frontend?:

No changes made to the `.vscode` folder.

## Changes

This ticket includes the following new changes:
- `@/shared/components`: Index barrel for easier importing.
- `@/shared/components/animations`: Added a reusable animation
component.
- `@/shared/components/auth`: Added a new folder in shared for
auth-specific components.
- `@/shared/content`: Global centralized static data.
- `@/pages/home`: Updated home to utilize centralized data-pool.
- `@/pages/contact`: Updated contact to remove hard-coded data and
utilize centralized data-pool.
- `@/pages/FAQ`: Updated faq to remove hard-coded data and utilize
centralized data-pool.
- `@/features/FAQ`: Updated FAQ hooks and data to utilize new
centralized data-pool.
- `@/layouts/navbar`: Updated navbar to utilize centralized data-pool.
- `@/layouts/footer`: Updated footer to utilize centralized data-pool.

## Does this use the v2 API?:

This PR does not use or change anything with the V2 api.

## Does this add a new communication between services?:

This PR does not add any new communications between services.