feat(VIST-CPC-1541): removed second implementation of reviews using the ownerId (#1281)
**JIRA:** [link to jira
ticket](https://champlainsaintlambert.atlassian.net/browse/CPC-1541)
## Context:
There was a second implementation of reviews where you needed to provide
the ownerId. This implementation was wrong because the ownerId is
supposed to be taken from the user your are currently using.
## Does this PR change the .vscode folder in petclinic-frontend?:
No

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
Commented out the second implementation of reviews.
Deleted the methods not used in the frontend
Changed the methods in the CustomerReviewsList to use the good api
Changed the permissions to make everyone able to make a review.
## Does this use the v2 API?:
No
## Does this add a new communication between services?:
No
## Before and After UI (Required for UI-impacting PRs)
No changes were made
## Dev notes (Optional)
N/A
## Linked pull requests (Optional)
N/A