feat(PLAT-CPC-1767): Refactor Go-Mailer To Follow Go Conventions (#1287)
**JIRA:**
[link](https://champlainsaintlambert.atlassian.net/browse/CPC-1767)

## Context:

This ticket is about following proper naming conventions when working
with go-lang. Go-lang has a very specific naming convention for
directories, files, and methods. I've simply implemented this naming
convention and applied it here.

Unfortunately, the old mailer-service was not following any naming
convention or typical go practices. The old mailer was doing very
java-style programming with impl's and such being present. This does not
follow nor represent proper go-lang conventions and this PR aims in
fixing this. In doing so, I've essentially re-made the entire service in
order to properly follow the go-conventions.

## Does this PR change the .vscode folder in petclinic-frontend?:

This pr does not change the `.vscode` directory.

## Changes

- `cmd/mailer/main.go`: This is the new main for the mailer service as
it follows the cmd/{service} convention for golang.
- `internal/http/handlers`: This is now the main handler for http
requests such as POST while using gin.
- `internal/http/middleware`: This is where the middlewares will be
located.
- `internal/mailer`: This is where the main mailer functionality is
located, for example a dialer constructor and the logic towards sending
an email from service.
- `internal/util`: This is where utilities such as env parsing / getting
are located.
- `pkg/mailer/types`: This is where the publicly available types are
(hence the usage of pkg).

The most important change, I moved the main.go situated in root to the
proper golang location which is in cmd/

## Does this use the v2 API?:

This does **not** involve the v2 API.

## Does this add a new communication between services?:

The communications between services has **not** changed.

## Before and After UI

No changes were made towards the front-end.

## Dev notes

I removed the old .NET code-base as it is completely useless.

---------

Co-authored-by: Jeremy Roos <149620614+JeremyNRoos@users.noreply.github.com>