Feat(AUTH-CPC-1700): Added mailer endpoints in billing service to allow the service to send emails (#1290)
**JIRA:**
[Link](https://champlainsaintlambert.atlassian.net/browse/CPC-1700?atlOrigin=eyJpIjoiZmM5M2M1OGNlMTkxNGRjZDg4MjMzMWNlMzMwNjdkZjAiLCJwIjoiaiJ9)
## Context:
Bills is trying to send an email when an owner pays a bill, so they
needed a way to use the mailer's email sending functions
## Does this PR change the .vscode folder in petclinic-frontend?:
no

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
- Added Mailer service client to bills to create communication between
services via "/mail" endpoint
- Added Auth service client to communicate with the auth service to
fetch the user details via a get method that takes a JWT token and a
user id. This returns a UserDetails object with the user's information
such as the email which is needed to send an email.
- Added a generateConfirmationEmail method that accepts a UserDetail
object which provides the user's email that is used to send an email to.
- Added tests for mailer service and auth service
- Updated C4 L2 Diagram to include connections from Billing-service to
Auth-service and Mailer-service
- Updated C4 L2 Diagram to remove connection from Visits service to
Mailer-service
- Added Code coverage for emailing method in Billing-service
## Does this use the v2 API?:
No versioning used
## Does this add a new communication between services?:
New connections Between Billing-service and:
- Mailer service
- Auth service
<img width="1329" height="1326" alt="image"
src="https://github.com/user-attachments/assets/5f1ba17a-24fc-44c8-bb26-0a90cfbda1ab"
/>

## Before and After UI (Required for UI-impacting PRs)
No changes to the UI
## Dev notes (Optional)
## Linked pull requests (Optional)

---------

Co-authored-by: isaacwallace123 <74162185+isaacwallace123@users.noreply.github.com>
Co-authored-by: Amir Ghadimi <sraldon24@outlook.com>