bug(AUTH-CPC-1769): Fix sign up bad request 400 (#1293)
**JIRA:**
[link](https://champlainsaintlambert.atlassian.net/browse/CPC-1769)

## Context:

This PR is to fix the sign up issue where email is not sent and bad
requets 400 appears when attempting to sign up. This bug was found
rather quickly in the logger dozzle from prod and then around an hour or
client-end docker compose testing took place in order to figure out the
issue.

## Does this PR change the .vscode folder in petclinic-frontend?:

No changes to the `.vscode` directory.

## Changes

- `pkg/mailer/types.go`: Changed the mapped types to align properly with
legacy go mailer service.

## Does this use the v2 API?:

This does not use or change V2 api.

## Does this add a new communication between services?:

This does not change the communication between services

## Before and After UI

There are no front-end changes

## Dev notes

I'm sorry for mistyping some types.

## Linked pull requests (Optional)

Pull request links