bug(INVT-CPC-1770) Fix Inventory Pages routing in the Angular that were leading to Product Pages. (#1299)
**JIRA:** https://champlainsaintlambert.atlassian.net/browse/CPC-1770
## Context:
Due to the product team refactoring their whole Angular routing from
shop-product to product, it caused some issues with the Inventory pages
routing due to the fact that some file.js were still using product as a
reference name, such as productDetails, productNew or productList. So
the issue was that we were not aware that shop-product would become
product only, so some pages were routing back to product pages in the
inventory code.

## Changes
What are the various changes and what other modules do those changes
affect.
- Remove duplicate module inventories ProductDetailsInfo
- productList is now inventoriesProductList
- productDetails is now inventoryProductDetails
- productNew is now inventoriesProductNew
- deleteProduct and deleteProductAll is now deleteInventoriesProduct and
deleteAllInventoriesProduct
- productNew is not inventoriesProductNew

Co-authored-by: OlivierGingras <149620440+IAMNOTBALAISE@users.noreply.github.com>