feat(VETS-CPC-1093): delete photo by vetId (#1307)
https://champlainsaintlambert.atlassian.net/browse/CPC-1093

Context:
The objective of that ticket is to make sur that only the admin and the
vets can delete a photo in the album or add it, and to have a
confirmation to make sure that you will delete the photo and you can't
undo that.

Does this PR change the .vscode folder in petclinic-frontend?:
Yes, I changed the frontend to make sure that it displays a message with
the vetid that you are in, and if you want to continue and delete the
photo in that section

Changes

-Front-end (vets)


Does this use the v2 API
Yes, exactly like all the other features

Does this add a new communication between services?:
No



Before

As you can see nothing appear and you can direclty delete the photo so
if you make a mistake your photo is gone forever


<img width="1470" height="956" alt="Screenshot 2025-10-14 at 5 21 09 PM"
src="https://github.com/user-attachments/assets/1fda31c5-0521-4cba-a9c0-c235f7d0d460"
/>


After

Now, when you click on delete, it asks you if you want to delete the
photo, and you have the option to continue or undo the action to keep
the photo. It displays the vet ID to make sure which vet those photos
are for so that you are sure that you are deleting your picture by
confirming your vet ID.

<img width="1470" height="956" alt="Screenshot 2025-10-14 at 5 22 20 PM"
src="https://github.com/user-attachments/assets/cbf63406-e223-4d4e-b9f8-a7db63a4c259"
/>