bug(CART-CPC-1777): Reverting changes for wrong API usage (#1314)
**JIRA:**

[Link](https://champlainsaintlambert.atlassian.net/browse/CPC-1777?atlOrigin=eyJpIjoiN2IwN2M3MTBhYWVmNDBmZGFiNjQ0Y2VlYzE3MjIzZmIiLCJwIjoiaiJ9)
## Context:
The ticket is changing getImage in the frontend to call v1 instead of v2
endpoint, reverting a change done in ticket CPC-1558. This change is
important because currently at the moment the PROD team plans on
deleting their v2 endpoints and are restructuring their frontend to
reflect those changes. If left unchanged, the call to v2 could break the
functionality of products.
Also, in the same ticket some calls were reverted back to v2 for carts
as well. This PR will be reverting those as well.
## Does this PR change the .vscode folder in petclinic-frontend?:
No
## Changes
- getImage calls the v1 API now
- Various calls that were working with v1 in relation to carts were
changed to use v2. Those have been reverted back to using the v1 API
## Does this use the v2 API?:
No
## Does this add a new communication between services?:
No
## Before and After UI (Required for UI-impacting PRs)
None

---------

Co-authored-by: Seddik Ayoub <102536731+Ayoub-Seddik@users.noreply.github.com>
Co-authored-by: Alejandro Marquez Vega <71571496+Chronyyx@users.noreply.github.com>