(feat/PROD-CPC-1659):Click product image navigates to product details… (#1345)
… page

**JIRA:**
https://champlainsaintlambert.atlassian.net/browse/CPC-1659?atlOrigin=eyJpIjoiMjkyNDU2NzU4M2I1NDc2N2JjZjA0YmYxNmMyNTA4MWQiLCJwIjoiaiJ9

## Context:
Images can also able to be clicked to navigate to product detail page

## Does this PR change the .vscode folder in petclinic-frontend?:
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
Product.tsx file - ImageContainer was moved inside clickable container
h2

## Does this use the v2 API?:
If the PR uses the v2 API, explain why

## Does this add a new communication between services?:
If the PR adds a new communication between services, make sure to update
the C4 L2.

## Before and After UI (Required for UI-impacting PRs)
If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.

## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links