Feat(PROD-CPC-1726): Create Api Call for Enums (#1369)
implemented a getProductsEnumValues api call in order to use for the
frontends instead of hardcoding them

**JIRA:** link to jira ticket
https://champlainsaintlambert.atlassian.net/browse/CPC-1726

## Context:
What is the ticket about and why are we doing this change.
For the angular Frontend, the enums had to be hardcoded, so to avoid
that, an Api to return them was created. the React FE should also use
this call

Be sure to include cgerard321 as a reviewer.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.
Implemented API call in product service, added it to the API gateway,
and created tests for it

## Does this use the v2 API?:
If the PR uses the v2 API, explain why
no
## Does this add a new communication between services?:
If the PR adds a new communication between services, make sure to update
the C4 L2.
no
## Before and After UI (Required for UI-impacting PRs)
If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.
<img width="665" height="286" alt="image"
src="https://github.com/user-attachments/assets/6f1f45c5-1e27-44d4-aeff-b1ccc2f1a1a3"
/>
<img width="650" height="447" alt="image"
src="https://github.com/user-attachments/assets/d3868c2c-5178-4a6c-9e78-edae35ca0eeb"
/>

## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links