Feat/CUST-CPC-1450_Get_Pet_Photo (#1377)
**JIRA:** link to jira ticket
https://champlainsaintlambert.atlassian.net/browse/CPC-1450

## Context:
What is the ticket about and why are we doing this change.

The ticket is about adding a GET profile picture for an owner's pets, so
an owner can customise their page with a picture of their lovely animals
and see it. We are doing this change to add customisation to the owner
profile and introduce more usages of the file service.

## Does this PR change the .vscode folder in petclinic-frontend?:
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.
NA
<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>
NA
## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.

The changes were mainly to the customers service reactive, the api
gateway and the petclinic front-end and targeted customer related files,
more precisely pet files to add the request param photo to the get pet
by id and fetch it in the file service's bucket. These changes affect
the previously mentioned 3 modules, for both their SRC and Test files.

## Does this use the v2 API?:
If the PR uses the v2 API, explain why
NA
## Does this add a new communication between services?:
If the PR adds a new communication between services, make sure to update
the C4 L2.
NA

## Before and After UI (Required for UI-impacting PRs)
If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.

<img width="1919" height="1079" alt="Capture d’écran 2025-10-23 170530"
src="https://github.com/user-attachments/assets/9953924e-e25b-4c23-8ed9-10d8c4fb06f5"
/>

<img width="1892" height="908" alt="Capture d’écran 2025-10-23 132804"
src="https://github.com/user-attachments/assets/887ea9d9-5d47-46a3-9cb5-a29f3fd119a0"
/>

## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links

---------

Co-authored-by: Mir Faiyazur Rahman <mirfaiyazrahman@gmail.com>
Co-authored-by: Gazou16 <146896335+Gazou16@users.noreply.github.com>