feat(CPC): Add dev tools (#1384)
**JIRA:** N/A

## Context:
I was bored and Rust looked fun. I also wanted to add a way for students
to see and manage the files in the database without having to deal with
exposing the Minio UI which was also recently changed and removed all
admin options.

Minio also does not seem to allow multiple users and RBAC. In the end I
want this tool ta have quick links to log viewer, dbgates, etc. This
will be tools to help see what happens in the deployed version. Ofc we
could always use already done solutions but where is the fun in that :)

## Changes
- Added a new Rust project that can communicate with the hosted bucket
so we can debug the images better.

## Before and After UI (Required for UI-impacting PRs)
UI will be coming one day.

## Dev notes (Optional)
I never coded in Rust before so this is probably poor code.
I did not add tests for now since I am still learning and the project
scope is still undefined.