CUST-CPC-1503-PetPhoto Default pictures (#1397)
**JIRA:** link to jira ticket
https://champlainsaintlambert.atlassian.net/browse/CPC-1503

## Context:
What is the ticket about and why are we doing this change.

The ticket is about allowing owners to see a default picture of a pet
based on a pet's petType if the photo for that pet is non-existent, has
a null field or if the file service is down, etc, then the default
picture gets displayed instead.

## Does this PR change the .vscode folder in petclinic-frontend?:
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.
NA
<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>
NA
## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.

The changes were made to the React Front-end to display a default
picture in case of failure of the file service or non-existent or
wrongful photoIds per pet type, and they are stored in the assets
folder. No other modules are affected, except the
customers-service-reactive where I added a default photoId to betty in
the DataSetUp for further testing.

## Does this use the v2 API?:
If the PR uses the v2 API, explain why
NA
## Does this add a new communication between services?:
If the PR adds a new communication between services, make sure to update
the C4 L2.
NA

## Before and After UI (Required for UI-impacting PRs)
If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.

<img width="1881" height="964" alt="Capture d’écran 2025-10-25 152110"
src="https://github.com/user-attachments/assets/96d1e39c-a9a0-45bc-a155-6aed7bb31bcd"
/>

<img width="1030" height="869" alt="Capture d’écran 2025-10-25 150730"
src="https://github.com/user-attachments/assets/72183f42-d483-4ca7-a7ec-8a21f7fde400"
/>

<img width="1013" height="862" alt="Capture d’écran 2025-10-25 150738"
src="https://github.com/user-attachments/assets/56b13a05-fdd8-41d2-9b86-2f4701bc2c2a"
/>

<img width="1886" height="948" alt="Capture d’écran 2025-10-25 145558"
src="https://github.com/user-attachments/assets/a4b7347a-25b1-4a59-ab82-b76cfed299f1"
/>


## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links