bug(INVT-CPC-1759): fix inventories on subsequent pages (#1419)
**JIRA:** https://champlainsaintlambert.atlassian.net/browse/CPC-1759

## Context:
In the inventories list, when inventories were added or moved to the
second page of the list, they would not be "rendered" in the queries.
This means that whenever you would search for an inventory it would not
be found due to it being on the 2nd page. This PR fixes that issue so
that all inventory queries (search, type, desc) now properly return the
found inventory regardless of the pagination.

## Does this PR change the .vscode folder in petclinic-frontend?:
No

## Changes
In `useSearchInventories.ts` I changed the filter behavior so searches
aggregate results from all backend pages instead of only the first page

## Does this use the v2 API?:
No

## Does this add a new communication between services?:
No

## Before and After UI (Required for UI-impacting PRs)
N/A